### 2023-02-05 #5129 @editorialbot
[PRE REVIEW]: SeqPanther: Sequence manipulation and mutation statistics toolset
**Submitting author:** @jsan4christ (<a href="http://orcid.org/0000-0002-5736-664X">James Emmanuel San</a>)
**Repository:** https://github.com/codemeleon/seqPanther
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
