+++ README.md
@@ -8,7 +8,7 @@ SeqPanther tool consists set of commands as follows:
 
 - **transform**: performs transformation of codon counter output in a format where the user can select a variant to integrate into the reference or assemblies using **nusub** command. The transformation can be performed after the removal of undesired changes in files generated by the codon counter or the first transformation is performed and undesired changes are removed from transformed data.
 
-**nucsub**: integrates the alteration list generated by **codoncounter** in a given reference or assembled genome based on the user recommendations. <!--Providing options to users to select changes of their interests-->
+- **nucsub**: integrates the alteration list generated by **codoncounter** in a given reference or assembled genome based on the user recommendations. <!--Providing options to users to select changes of their interests-->
 
 ## Operating system compatibility
 
@@ -96,10 +96,10 @@ Recursive use of newly generated consensus sequences might result in incorrect f
 
 # Bug reporting
 
-Please report bugs on this portal or associated GitHub repositories of the command
+For bug reporting, please open an issue.
 
-# Cite
+# Citation
 
-If you use this tool, please cite it as
+If you use this software please cite:
 
-Bib for the opensource software journal (JOSS)
+SeqPanther: Sequence manipulation and mutation statistics toolset. James Emmanuel San, Stephanie Van Wyk, Houriiyah Tegally, Simeon Eche, Eduan Wilkinson, Aquillah M. Kanzi, Tulio de Oliveira, Anmol M. Kiran. bioRxiv 2023. doi: [10.1101/2023.01.26.525629](https://doi.org/10.1101/2023.01.26.525629)
