### 2023-02-05 #5129 @jsan4christ
Thank you,

I have gone through the list of reviews and here are some of the
outstanding profiles. I don't know any of them but hope that they would be
great candidates.

https://github.com/ctb
https://github.com/cbrueffer
https://github.com/serine
https://github.com/a-slide

On Sun, Feb 5, 2023 at 10:52 AM The Open Journals editorial robot <
***@***.***> wrote:

> *Submitting author:* @jsan4christ <https://github.com/jsan4christ> (James
> Emmanuel San <http://orcid.org/0000-0002-5736-664X>)
> *Repository:* https://github.com/codemeleon/seqPanther
> *Branch with paper.md* (empty if default branch):
> *Version:* v1.0.0
> *Editor:* Pending
> *Reviewers:* Pending
> *Managing EiC:* Kevin M. Moerman
> Status
>
> [image: status]
> <https://joss.theoj.org/papers/1cd491667547329b5c7841df9d8506f5>
>
> Status badge code:
>
>
> *Author instructions*
>
> Thanks for submitting your paper to JOSS @jsan4christ
> <https://github.com/jsan4christ>. *Currently, there isn't a JOSS editor
> assigned* to your paper.
>
> @jsan4christ <https://github.com/jsan4christ> if you have any suggestions
> for potential reviewers then please mention them here in this thread
> (without tagging them with an @). In addition, this list of people
> <https://bit.ly/joss-reviewers> have already agreed to review for JOSS
> and may be suitable for this submission (please start at the bottom of the
> list).
>
> *Editor instructions*
>
> The JOSS submission bot @editorialbot <https://github.com/editorialbot>
> is here to help you find and assign reviewers and start the main review. To
> find out what @editorialbot <https://github.com/editorialbot> can do for
> you type:
>
> @editorialbot commands
>
