+++ seqPanther/seqPatcher/seqpatcher.py
@@ -203,7 +203,6 @@ def ranges(lst, given_gap=0):
     A generator returns list of range based on given numbers
     [1,2,3,4,5, 10, 11, 12, 13, 14] => [[1,5], [10,14]]
     """
-    # print(given_gap, "Anmol")
     lst_sorted = sorted(lst)
     init = 0
     for num in range(1, len(lst_sorted)):
@@ -222,7 +221,6 @@ def useful_range(lst, gap=10):  # TODO: Add in command
     # or they want other fragment to be selected
 
     trange = list(ranges(lst, gap))
-    # print(trange, "kiran")
     pre_frag_len = 0
     myrange = trange[0]
     for rng in trange:
@@ -449,8 +447,6 @@ def merge_base_peak(nuc_df, peak_dict):
 
     nuc_df["idx"] = list(nuc_df.index)
 
-    # df = enumerate_columns(df)
-    # print(df)
     to_drop = ["idx"]
     for nuc in peak_dict:
         # Expecting sequence will not change while aligning, except insertion of gaps
@@ -506,8 +502,6 @@ def rep_paired_base(lst, ambiguous=False):
                         base = bs
                 return base
         else:
-            # if ambiguous:
-            # print(lst["F_peak"], lst["R_peak"], "Kkiran", lst)
             amb_base = tuple(list(set(lst["F_peak"]) & set(lst["R_peak"])))
             if not amb_base:
                 amb_base = tuple(list(set(lst["F_peak"]) | set(lst["R_peak"])))
@@ -582,8 +576,6 @@ def aln_clean(aln_df, gap=15, ambiguous=False):  # , at, res_fold):
                         else:
                             aln_df.loc[ltc:, rev[dir]] = "-"
 
-        # print(aln_df.loc[u_range_r[-1] + 1 : u_range_r[-1] + 30, "R"])
-        # print(u_range_f, u_range_r, "dead zone")
         u_range = [
             np.min([u_range_f[0], u_range_r[0]]),
             np.max([u_range_f[-1], u_range_r[-1]]),
@@ -591,9 +583,6 @@ def aln_clean(aln_df, gap=15, ambiguous=False):  # , at, res_fold):
         aln_df.loc[:u_range[0] - 1, ["F", "R"]] = "-"
         aln_df.loc[u_range[1] + 1:, ["F", "R"]] = "-"
         sang_type = "P"  # For paired
-        # print(aln_df.head())
-        # TODO: Must check in case of mismatch and ambigious it remove it carefully
-        # print(idx)
     else:
         if "F" in aln_df:
             idx = aln_df[aln_df["F"] != "-"].index
@@ -616,21 +605,16 @@ def aln_clean(aln_df, gap=15, ambiguous=False):  # , at, res_fold):
         ambi_indexes = aln_df[aln_df[col].isin(list("NRYKMSWBDHV"))].index
         ambi_indexes = ambi_indexes[(ambi_indexes >= u_range[0])
                                     & (ambi_indexes <= u_range[-1])]
-        # print(u_range, ambi_indexes, "anmol", u_range)
 
         if len(ambi_indexes):
-            print(ambi_indexes)
             ambi_index_left = ambi_indexes[ambi_indexes < u_range[0] + 20]
             ambi_index_right = ambi_indexes[ambi_indexes > u_range[-1] - 20]
             if ambi_index_left.any():
                 u_range[0] = ambi_index_left.max()
             if ambi_index_right.any():
                 u_range[-1] = ambi_index_right.min()
-            # print(u_range, "testing")
-            # exit()
             aln_df.loc[:u_range[0] - 1, col] = "-"
             aln_df.loc[u_range[1] + 1:, col] = "-"
[…]
