### 2023-02-05 323c984
Muscle version correction and example for seqpatcher
