+++ seqPanther/seqPatcher/README.md
@@ -1,0 +1,178 @@
+# SARS-CoV-2 Sequencing: Merge Sanger
+
+The script seqpatcher.py integrates the nucleotides sequenced with the Sanger platform into the corresponding HTS-based (**SARS-CoV-2**) genome assembly.
+
+<!--The script seqpatcher.py integrates the Sanger sequenced segment into the corresponding HTS based assembled incomplete genomes.-->
+
+# Dependencies
+
+## External tools (non-Python)
+
+- [MUSCLE](https://www.drive5.com/muscle/downloads.htm)
+  - To perform multiple sequence alignment
+- [BLAT](https://hgdownload.soe.ucsc.edu/admin/exe/)
+  - To query sequence location in the genome
+
+Both tools can either be downloaded from the links provided, or installed using Anaconda or Ubuntu Linux repositories. Tools must be in the system path and accessible via the terminal.
+
+### Anaconda command
+
+```bash
+conda install blat muscle
+```
+
+### Ubuntu command
+
+Note: Only system administrators are allowed.
+
+```bash
+sudo apt install blat muscle
+```
+
+## Python (v3.6+) modules
+
+- pandas==1.2.4
+- numpy==1.19.1
+- biopython==1.78
+- click==7.1.2
+
+### Installation
+
+- Install dependencies with `pip install -r requirements.txt`, download `seqpatcher.py` to your working directory and execute it as instructed in the [Execution](#execution) section.
+- Install with `pip install git+https://github.com/krisp-kwazulu-natal/seqPatcher` or `git clone https://github.com/krisp-kwazulu-natal/seqPatcher && cd seqPatcher && python setup.py install`. `seqpatcher.py` should be installed in the system path and accessible from the command-line in the terminal.
+
+## File and sequence naming convention.
+
+1. File name or sample name must not have special characters.
+2. Assembly files should be written as <sample_name>.fasta. Sequence's name should be identical to <sample_name>.
+3. If assembly files contain multiple sequences, assembly file is generated for each sample ID as explained in point 2.
+4. If all the assemblies are in one file, name can be anything with fasta extension. However, the sequence names must match <sample_name>.
+5. ab1 file should be named as <sample_name>.\<xyz>.ab1. Where xyz can be any string.
+6. If user already has preprocessed Sanger trace files to fasta, it should be named as <sample_name>.\<xyz>.fasta. Where xyz can be any character string.
+
+<a name='cmdoptions'>
+
+# Commandline options
+
+```
+Usage: seqpatcher.py [OPTIONS]
+
+  Reports nucleotide sequence from Sanger chromatogram data based on user
+  provided parameters and integrate that in assembly generated using NGS
+  data
+
+Options:
+  -s, --sanger-ab1 TEXT           Folder containing Sanger sequencing trace
+                                  files or Fasta files generated from trace
+                                  files.  [default: sanger_ab1]
+
+  -a, -assemblies-foder TEXT      Folder containing HTS generate incomplete
+                                  assembly Fasta files  [default: assemblies]
+
+  -o, --out-dir TEXT              Result output Folder  [default: Results]
+  -t, --tab TEXT                  CSV file for overlapping assemblies and
+                                  Sanger IDs. If not specified, stdout.
+
+  -O, --output-fasta TEXT         Ouput file name for Fasta from sanger ab1
+  -R, --ref-gene-fasta-file TEXT  Refence gene in fasta format
+  -c, --clean-intermediate BOOLEAN
+                                  Remove intermediate files  [default: True]
+  -g, --gap-allowed INTEGER       Minimum gap length between aligned fragment
+                                  to consider the alignment continuous
+                                  [default: 10]
+
+  -3, --only-3-nuc BOOLEAN        Allow multiple 3 nucleotide InDels else
+                                  replace with reference nucleotides or Ns
[…]
