### 2023-02-09 #5129 @Kevin-Mattheus-Moerman
@jsan4christ thanks for this submission. I am the AEiC for this track and here to help process initial steps and to check if this work is in scope for JOSS. I have just triggered a scope review given the small size of this submission (see also: https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). The scope review will take about 2 weeks to complete.
