### 2023-02-10 7ad3109
Plot range and max depth
