+++ seqPanther/CodonCounter/coors_with_changes.py
@@ -1,9 +1,10 @@
 #!/usr/bin/env python
 import pysam
 import pandas as pd
+import numpy as np
 
 import pyfaidx
-from os import path
+from os import path, system
 
 from .subs import sub_table
 from .indel_frames import indel_frames
@@ -11,7 +12,9 @@ from .indel_frames import indel_frames
 
 def changed_coordinates(params, bam):
     print(f"Analysing {bam}.")
+    ref = params["ref"]
     rid = params["rid"]
+    tmp_dir = params["tmp_dir"]
     start = params["start"]
     end = params["end"]
     endlen = params["endlen"]
@@ -20,6 +23,7 @@ def changed_coordinates(params, bam):
     min_mapping_quality = params["min_mapping_quality"]
     min_base_quality = params["min_base_quality"]
     min_seq_depth = params["min_seq_depth"]
+    max_seq_depth = params["max_seq_depth"]
     alt_nuc_count = params["alt_nuc_count"]
     ignore_overlaps = params["ignore_overlaps"]
 
@@ -30,73 +34,114 @@ def changed_coordinates(params, bam):
         print(samfile.references)
         print(f"Ignoring {bam}")
         return
-    iter = samfile.pileup(
-        rid,
-        start,
-        end,
-        ignore_orphans=ignore_orphans,
-        min_base_quality=min_base_quality,
-        min_mapping_quality=min_mapping_quality,
-        ignore_overlaps=ignore_overlaps,
-        max_depth=1000000,
-    )
+
+    vcf_file = f'{tmp_dir}/{path.split(bam)[-1].split(".")[0]}.vcf'
+
+    # command = f"bcftools mpileup -x -d 20000000 -m 1 -e 10 -L 10000000 --open-prob 10  -Q 0 -A -B -C0 --annotate FORMAT/AD  -Ob -r  NC_045512.2:20000-25000 --no-BAQ -f NC_045512.2.fasta -o test.bcf K032298-consensus_alignment_sorted.bam"
+    command = f"bcftools mpileup --annotate FORMAT/AD -d {max_seq_depth} -L {max_seq_depth} -m 1 -e 40 --open-prob 40 -B -q {min_mapping_quality} -C0 -Q {min_base_quality} -r {rid}:{start}-{end} --no-BAQ -f {ref} -o {vcf_file} {bam}"
+    if ignore_overlaps:
+        command += ' -x'
+    if ignore_orphans:
+        command += ' -A'
+
+    command += " 2>/dev/null"
+
+    system(command)
+    vcf = pd.read_table(vcf_file,
+                        header=None,
+                        comment="#",
+                        usecols=[0, 1, 7, 9])
+    vcf[9] = vcf[9].apply(lambda x: x.split(':')[1])
+    vcf[9] = vcf[9].apply(lambda x: np.array(list(map(int, x.split(',')))))
+    vcf['total'] = vcf[9].apply(lambda x: np.sum(x))
+    depth = vcf[[1, 'total']].rename(columns={
+        1: 'coor',
+        'total': 'depth'
+    }).sort_values('depth',
+                   ascending=False).drop_duplicates('coor').sort_values('coor')
+
+    vcf = vcf[vcf['total'] > min_seq_depth]
+    vcf[9] = vcf[9] / vcf['total']  # TODO: Keep a copy of this for plotting
+    vcf[9] = vcf[9].apply(lambda x: x[1:])
+
+    vcf = vcf[vcf[9].apply(lambda x: np.sum(x > alt_nuc_count)) > 0]
+    print(vcf.shape)
+
     coordinates_with_change = {}
     indel_pos_type_size = {"coor": [], "indel": [], "seq": []}
[…]
