### 2023-02-14 #5129 @Kevin-Mattheus-Moerman
@jsan4christ on the above paper compilation error you need to use quotes if you have multiple affiliations, e.g. `1,2` should be written `"1,2"`. I have fixed this for you here: https://github.com/codemeleon/seqPanther/pull/1.
