### 2023-02-20 e99aaf8
Preparation for next version of codon counter
