+++ seqPanther/CodonCounter/coors_with_changes.py
@@ -65,15 +65,21 @@ def changed_coordinates(params, bam):
     vcf[9] = vcf[9].apply(lambda x: x[1:])
 
     vcf = vcf[vcf[9].apply(lambda x: np.sum(x > alt_nuc_count)) > 0]
+    # print(vcf)
+    vcf = vcf[vcf[1] == 21764]
 
     coordinates_with_change = {}
-    # indel_pos_type_size = {"coor": [], "indel": [], "seq": []}
     indel_pos_type_size = {}
-    already_used_coordinates = []
+    vcf = vcf[vcf[7].str.startswith('INDEL')].head(1)
     for row in vcf.to_dict('records'):
-        start, end = (row[1] - 2, row[1] +
-                      3) if row[7].startswith('INDEL') else (row[1] - 1,
-                                                             row[1])
+        is_indel = False
+        if row[7].startswith('INDEL'):
+            is_indel = True
+        start, end = (row[1] - 1, row[1])
+        if not is_indel:
+            continue
+        print(start, end)
+        # print(coordinates_with_change)
         iter = samfile.pileup(
             rid,
             start,
@@ -84,90 +90,114 @@ def changed_coordinates(params, bam):
             ignore_overlaps=ignore_overlaps,
             max_depth=max_seq_depth,
         )
+        bases = {}
+        read_depth = 0
+        reads_to_remove = {}
         for pileupcol in iter:
-            if pileupcol.pos >= end:
-                break
-            if ((pileupcol.pos < start)
-                    or (pileupcol.pos in already_used_coordinates)
-                    or (pileupcol.n < min_seq_depth)):
+            # if pileupcol.pos < start - 3:
+            # continue
+            # print(pileupcol.pos, pileupcol.n)
+            if pileupcol.pos != start:
                 continue
-            if (pileupcol.pos >= start) & (pileupcol.pos < end):
-                already_used_coordinates.append(pileupcol.pos)
-                # TODO: Include base call quality
-                bases = {}
-
-                for pread in pileupcol.pileups:
-
+            read_depth = pileupcol.n
+            reads_to_remove_count = {}
+            for pread in pileupcol.pileups:
+                if not pread.query_position:
+                    continue
+                add_left = pread.query_position - pread.alignment.query_alignment_start
+                add_right = pread.alignment.query_alignment_end - pread.query_position
+                if is_indel:
                     if pread.indel and pread.indel % 3 == 0:
-                        indel_seq = ''  # Deletion
-                        if pread.pos not in indel_pos_type_size:
-                            indel_pos_type_size[pileupcol.pos] = {}
-
-                        indel_pos_type_size["coor"].append(pileupcol.pos)
-                        indel_pos_type_size["indel"].append(pread.indel)
-                        if pread.indel > 0:
-                            indel_seq = pread.alignment.query_sequence[
-                                pread.query_position + 1:pread.query_position +
-                                1 + pread.indel]
-
-                            # indel_pos_type_size["seq"].append()
-                        continue
-
-                    pread_len = len(pread.alignment.query_sequence)
-                    if not pread.is_del and not pread.is_refskip:
-                        if (pread.query_position < endlen or
-                            (pread_len - pread.query_position + 1) < endlen):
-                            continue
-                        tbase = pread.alignment.query_sequence[
-                            pread.query_position]
-                        if tbase not in bases:
-                            bases[tbase] = {
-                                "nuc_count": 0,
-                                "codon_count": {},
-                            }
-                        bases[tbase]["nuc_count"] += 1
-                        add_left = 2 - pread.query_position
+                        add_left = 3 - add_left
                         add_left = add_left if add_left > 0 else 0
[…]
