+++ seqPanther/CodonCounter/indel_frames.py
@@ -7,18 +7,14 @@ from os import path
 
 
 def indel_frames(indel_pos_type_size, bam, params):
-    rid = params["rid"]
     gff_data = params["gff_data"]
-    alt_nuc_count = params["alt_nuc_count"]
     alt_codon_frac = params["alt_codon_frac"]
-    sample = path.split(bam)[1].split(".bam")[0]
-
-    deletion_frame = {}
-    insertion_frame = {}
-
+    rid = params["rid"]
+    sample = params["sample"]
     coors = set(indel_pos_type_size["coor"])
     indel_pos_type_size["amino_pos"] = 0
     indel_pos_type_size["codon_pos"] = 0
+    shift, r_shift = 0, 0
 
     for coor in coors:
         # TODO: use df.to_dict('records'). more detail https://towardsdatascience.com/heres-the-most-efficient-way-to-iterate-through-your-pandas-dataframe-4dad88ac92ee
@@ -40,7 +36,6 @@ def indel_frames(indel_pos_type_size, bam, params):
                                         indel_pos_type_size["coor"] == coor,
                                         ["ref", "read"]].applymap(lambda x: x[
                                             3 - shift:-(3 - r_shift)])
-            print(indel_pos_type_size)
 
             amino_pos = (adjusted_coor -
                          gff_row["start"]) // 3  # - 1 * row["indel"]
@@ -67,23 +62,67 @@ def indel_frames(indel_pos_type_size, bam, params):
                                     "amino_pos"] = amino_pos
             indel_pos_type_size.loc[indel_pos_type_size["coor"] == coor,
                                     "codon_pos"] = codon_pos
+    cols = indel_pos_type_size.columns.tolist()
+    cols.remove('count')
+    indel_pos_type_size = indel_pos_type_size.groupby(
+        cols)['count'].sum().reset_index()
+
+    indel_pos_type_size_ref_support = indel_pos_type_size.groupby([
+        'coor', 'depth', 'indel', 'ref', 'amino_pos', 'codon_pos'
+    ])['count'].sum().reset_index().rename(columns={'count': 'ref_count'})
+    indel_pos_type_size = indel_pos_type_size.merge(
+        indel_pos_type_size_ref_support,
+        on=['coor', 'depth', 'indel', 'ref', 'amino_pos', 'codon_pos'])
+    del indel_pos_type_size_ref_support
+    indel_pos_type_size['ref_count'] = indel_pos_type_size[
+        'depth'] - indel_pos_type_size['ref_count']
+    indel_pos_type_size = indel_pos_type_size[
+        indel_pos_type_size['count'] > indel_pos_type_size['depth'] *
+        alt_codon_frac]
+
+    if not len(indel_pos_type_size):
+        return pd.DataFrame()  # TODO: Add dummy columns
+
+    indels_changes = indel_pos_type_size.copy()
+    indels_changes["coor"] = indels_changes['codon_pos'] + shift
+    # indels_changes.apply(
+    # lambda x: 2 if len(x['ref']) > len(x['read']) else 1, axis=1)
+
+    indels_changes["Nucleotide Percent"] = indels_changes.apply(
+        lambda x: x['count'] / x['depth'], axis=1)
+
+    indels_changes["Nucleotide Frequency"] = indels_changes.apply(
+        lambda x: 'del' + x['ref'][r_shift:-shift] if (len(x['ref']) > len(x[
+            'read'])) else 'ins' + x['read'][r_shift:-shift],
+        axis=1)
+
+    print(indels_changes)
+
[…]

[…]
