+++ docs/CONTRIBUTING.md
@@ -1,0 +1,5 @@
+#Contribute to codemeleon/seqPanther
+
+Make your first contribution to this repository by tackling one of the issues listed below.
+
+Each issue displayed here is a "good first issue," selected for its relative approachability for first-time contributors.
