### 2023-02-28 #5129 @jsan4christ
Thank you for quality checkes. I have added the quotes @editorialbot generate pdf. The ones without doi's are web pages and therefore have not doi's but just the URL's @editorial bot check references. I have added the CONTRIBUTING.md file under docs/CONTRIBUTING.md. Let me know if that is okay.
