+++ seqPanther/CodonCounter/coors_with_changes.py
@@ -43,7 +43,6 @@ def changed_coordinates(params, bam):
         command += ' -x'
     if not ignore_orphans:
         command += ' -A'
-    print(command)
 
     command += " 2>/dev/null"
 
@@ -66,17 +65,25 @@ def changed_coordinates(params, bam):
     vcf[9] = vcf[9].apply(lambda x: x[1:])
 
     vcf = vcf[vcf[9].apply(lambda x: np.sum(x > alt_nuc_count)) > 0]
+    vcf = vcf.sort_values(1, ascending=False)
+    # TODO: If vcf has no change, return empty dataframe
     # print(vcf)
     # vcf = vcf[vcf[1] == 21764]
+    coors = set(vcf[1] - 1)
+    print(vcf)
 
     coordinates_with_change = {}
     indel_pos_type_size = {}
     reads_to_remove = {}
-    for row in vcf.to_dict('records'):
-        is_indel = False
-        if row[7].startswith('INDEL'):
-            is_indel = True
-        start, end = (row[1] - 1, row[1])
+    already_seen = set()
+    for row in vcf.itertuples():
+        # print(row)
+        # is_indel = False
+        # if row[3].startswith('INDEL'):
+        # is_indel = True
+        start, end = (row[2] - 1, row[2])
+        if start in already_seen:
+            continue
         # if not is_indel:
         # continue
         iter = samfile.pileup(
@@ -92,11 +99,13 @@ def changed_coordinates(params, bam):
         bases = {}
         read_depth = 0
         for pileupcol in iter:
-            # if pileupcol.pos < start - 3:
+            if (pileupcol.pos not in coors) or (pileupcol.pos in already_seen):
+                continue
+            if pileupcol.pos in coors:
+                already_seen.add(pileupcol.pos)
+
             # continue
             # print(pileupcol.pos, pileupcol.n)
-            if pileupcol.pos != start:
-                continue
             read_depth = pileupcol.n
             reads_to_remove_count = {}
             for pread in pileupcol.pileups:
@@ -104,37 +113,36 @@ def changed_coordinates(params, bam):
                     continue
                 add_left = pread.query_position - pread.alignment.query_alignment_start
                 add_right = pread.alignment.query_alignment_end - pread.query_position
-                if is_indel:
-                    if pread.indel and pread.indel % 3 == 0:
-                        add_left = 3 - add_left
-                        add_left = add_left if add_left > 0 else 0
-                        overhang = add_right  # to remove sequence which terminate in between
-                        add_right = 4 - add_right
-                        add_right = add_right if add_right > 0 else 0
+                if pread.indel and pread.indel % 3 == 0:
+                    add_left = 3 - add_left
+                    add_left = add_left if add_left > 0 else 0
+                    overhang = add_right  # to remove sequence which terminate in between
+                    add_right = 4 - add_right
+                    add_right = add_right if add_right > 0 else 0
 
-                        if pread.indel < 0:
-                            if overhang <= abs(pread.indel):
-                                if overhang not in reads_to_remove_count:
-                                    reads_to_remove_count[overhang] = 0
-                                reads_to_remove_count[overhang] += 1
+                    if pread.indel < 0:
+                        if overhang <= abs(pread.indel):
+                            if overhang not in reads_to_remove_count:
+                                reads_to_remove_count[overhang] = 0
+                            reads_to_remove_count[overhang] += 1
 
-                            ref_sq = sequences[start + 1 - 3:start + 1 + 3 -
-                                               pread.indel].seq
-                            read_sq = '-' * add_left + pread.alignment.query_sequence[
[…]
