+++ seqPanther/CodonCounter/subs.py
@@ -7,7 +7,7 @@ from .codon_table import codon_table
 from Bio import Seq
 
 
-def sub_table(coordinates_with_change, params):
+def sub_table(coordinates_with_changes, params):
     sequences = params["sequences"]
     rid = params["rid"]
     sample = params["sample"]
@@ -16,7 +16,8 @@ def sub_table(coordinates_with_change, params):
 
     gff_data = params["gff_data"]
     min_seq_depth = params["min_seq_depth"]
-    keys = set(coordinates_with_change)
+    keys = set(coordinates_with_changes)
+    coordinates_with_change_cds = {}
 
     for (
             _,
@@ -24,11 +25,11 @@ def sub_table(coordinates_with_change, params):
     ) in gff_data.iterrows():  # TODO: Change it to itertuple for large gff
         selected_coordinates = keys & set(range(row["start"], row["end"] + 2))
         for selected_coordinate in selected_coordinates:
-            coordinates_with_change[selected_coordinate]["start"] = row[
-                "start"]
-            coordinates_with_change[selected_coordinate]["end"] = row["end"]
-            coordinates_with_change[selected_coordinate]["strand"] = row[
-                "strand"]
+            coordinates_with_change = coordinates_with_changes[
+                selected_coordinate].copy()
+            coordinates_with_change["start"] = row["start"]
+            coordinates_with_change["end"] = row["end"]
+            coordinates_with_change["strand"] = row["strand"]
             shift = (selected_coordinate - row["start"]) % 3
             ref_base = sequences[selected_coordinate].seq
             ref_codon = sequences[selected_coordinate -
@@ -36,13 +37,19 @@ def sub_table(coordinates_with_change, params):
             ref_codon_count = 0
             total_codon_count = 0
             codon_counts = {}
-            bases = coordinates_with_change[selected_coordinate]["bases"]
+            bases = coordinates_with_change["bases"]
+
+            if selected_coordinate == 28460:
+                print(bases)
             for base in bases.keys():
 
                 codons = bases[base]["codon_count"]
                 local_codons = {}
+
                 for extended_codon in codons.keys():
                     codon = extended_codon[2 - shift:5 - shift]
+                    if selected_coordinate == 28460:
+                        print(extended_codon, codon, shift, codons)
                     if '-' in codon:
                         continue
                     if codon not in local_codons:
@@ -57,12 +64,11 @@ def sub_table(coordinates_with_change, params):
                     ref_codon_count += local_codons[ref_codon]
                 except KeyError:
                     pass
-                coordinates_with_change[selected_coordinate]["bases"][base][
+                coordinates_with_change["bases"][base][
                     "codon_count"] = local_codons
 
             # NOTE: Removing less less common codons
-            coordinates_with_change[selected_coordinate][
-                "total_codon_count"] = total_codon_count
+            coordinates_with_change["total_codon_count"] = total_codon_count
 
             codons_to_delete = []
             for codon in codon_counts.keys():
@@ -72,8 +78,8 @@ def sub_table(coordinates_with_change, params):
             for base in list(bases.keys()):
                 for codon in codons_to_delete:
                     try:
-                        del coordinates_with_change[selected_coordinate][
-                            "bases"][base]["codon_count"][codon]
+                        del coordinates_with_change["bases"][base][
+                            "codon_count"][codon]
                     except KeyError:
                         pass
 
@@ -81,33 +87,30 @@ def sub_table(coordinates_with_change, params):
             if row["strand"] == "-":
                 ref_codon = str(Seq.Seq(ref_codon).reverse_complement())
                 ref_base = str(Seq.Seq(ref_base).reverse_complement())
-                for k in coordinates_with_change[selected_coordinate]["bases"]:
[…]
