### 2023-03-07 3215e4c
Ignoring codons with Ns
