+++ README.md
@@ -1,29 +1,33 @@
 # SeqPanther
 
-SeqPanther tool consists set of commands as follows:
+`SeqPanther` is a Python application that provides the user with a suite of tools to further interrogate the circumstance under which certain mutations occur or are missed and to enables the user to modify the consensus as needed. The tool is applicable to non-segmented bacterial and viral genomes where reads are mapped to a reference sequence. `SeqPanther` generates detailed reports of mutations identified within a genomic segment or positions of interest, including visualization of the genome coverage and depth. The tool is particularly useful in the examination of multiple next-generation sequencing (NGS) short-read samples. Additionally, we have integrated `Seqpatcher`[@Singh_2022] which supports the merging of Sanger sequences, or consensus thereof into their respective NGS consensus.
 
-- **seqpatcher**: integrates sanger sequencing of missing regions of an incomplete assembly to the assembly. This command is a modification of the [SeqPatcher tool](https://github.com/krisp-kwazulu-natal/seqPatcher).
+SeqPanther consists of the following set of commands:
 
 - **codoncounter**: performs variant calling and generates nucleotide stats at variant sites and reports the impacts of nucleotide changes on amino acids in the translated proteins.
 
-- **transform**: performs transformation of codon counter output in a format where the user can select a variant to integrate into the reference or assemblies using **nusub** command. The transformation can be performed after the removal of undesired changes in files generated by the codon counter or the first transformation is performed and undesired changes are removed from transformed data.
+- **cc2ns** (codon counter 2 nucleotide substitution): performs transformation of codon counter output in a format where the user can select a variant to integrate into the reference or assemblies using the **cc2ns** command. The user is encourage to inspect and edit the output from **codoncounter** to only keep the desired changes before running the **cc2ns** command.
 
-- **nucsub**: integrates the alteration list generated by **codoncounter** in a given reference or assembled genome based on the user recommendations. <!--Providing options to users to select changes of their interests-->
+- **cc2ns**: accepts as input the changes file generated by **cc2ns** and modifies the consensus sequences accordingly i.e. adds or changes the mutations as specified in the file. <!--Providing options to users to select changes of their interests-->
+
+- **seqpatcher**: integrates sanger sequencing of missing regions of an incomplete assembly to the assembly. This command is a modification of the [SeqPatcher tool](https://github.com/krisp-kwazulu-natal/seqPatcher).
+
+
 
 ## Operating system compatibility
 
-Unix platforms.
+Unix and OSX Commandline Application. 
 
 ## Dependencies
 
-The tool relies on multiple external programs and python modules as listed below:
+The tool relies on multiple external opensource programs and python modules as listed below:
 
 ### External Tools
 
 1. Samtools
 
    - For sorting bam files and indexing them.
-   - `conda install samtools` to install.
+   - `conda install samtools`
 
 2. Bcftools
 
@@ -49,23 +53,29 @@ The tool relies on multiple external programs and python modules as listed below
 
 ## Manual installation
 
-1. `git clone https://github.com/codemeleon/seqPanther`
-2. `cd seqPanther`
-3. `python setup.py install`
[…]
