+++ README.md
@@ -1,6 +1,6 @@
 # SeqPanther
 
-`SeqPanther` is a Python application that provides the user with a suite of tools to further interrogate the circumstance under which certain mutations occur or are missed and to enables the user to modify the consensus as needed. The tool is applicable to non-segmented bacterial and viral genomes where reads are mapped to a reference sequence. `SeqPanther` generates detailed reports of mutations identified within a genomic segment or positions of interest, including visualization of the genome coverage and depth. The tool is particularly useful in the examination of multiple next-generation sequencing (NGS) short-read samples. Additionally, we have integrated `Seqpatcher`[@Singh_2022] which supports the merging of Sanger sequences, or consensus thereof into their respective NGS consensus.
+`SeqPanther` is a Python application that provides the user with a suite of tools to further interrogate the circumstance under which certain mutations occur or are missed and to enables the user to modify the consensus as needed. The tool is applicable to non-segmented bacterial and viral genomes where reads are mapped to a reference sequence. `SeqPanther` generates detailed reports of mutations identified within a genomic segment or positions of interest, including visualization of the genome coverage and depth. The tool is particularly useful in the examination of multiple next-generation sequencing (NGS) short-read samples. Additionally, we have integrated `Seqpatcher`[singh et. al.](https://www.microbiologyresearch.org/content/journal/mgen/10.1099/mgen.0.000774) which supports the merging of Sanger sequences, or consensus thereof into their respective NGS consensus.
 
 SeqPanther consists of the following set of commands:
 
@@ -11,8 +11,6 @@ SeqPanther consists of the following set of commands:
 - **cc2ns**: accepts as input the changes file generated by **cc2ns** and modifies the consensus sequences accordingly i.e. adds or changes the mutations as specified in the file. <!--Providing options to users to select changes of their interests-->
 
 - **seqpatcher**: integrates sanger sequencing of missing regions of an incomplete assembly to the assembly. This command is a modification of the [SeqPatcher tool](https://github.com/krisp-kwazulu-natal/seqPatcher).
-
-
 
 ## Operating system compatibility
 
