+++ README.md
@@ -85,8 +85,6 @@ This command help is accessible at `seqpanther seqpatcher` or `seqpanther seqpat
 
 # Example
 
-
-
 1. A sample dataset of five BAM files from South African SARS-CoV-2 samples can be downloaded from [SeqPanther zenodo](https://zenodo.org/record/7601513) page. The BAM files were generated by mapping reads mapped against reference (NC_045512.2) using [BWA](https://github.com/lh3/bwa). BWA was also used to generate the SAM files which were converted to BAM and sorted using samtools.
 
 2. Reference sequences in Fasta format and genome annotation GFF files were downloaded from [NCBI Genome Fasta](https://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/009/858/895/GCF_009858895.2_ASM985889v3/GCF_009858895.2_ASM985889v3_genomic.fna.gz) and [ NCBI Genome GFF](https://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/009/858/895/GCF_009858895.2_ASM985889v3/GCF_009858895.2_ASM985889v3_genomic.gff.gz) respectively. The downloaded files need to un-archived.
@@ -111,17 +109,13 @@ The command will generate four outputs in the current folder including: `sub_out
 
 11. To run `seqpanther seqpatcher` on the example data, use `seqpanther seqpatcher -s examples/seqpatcher/ab1 -a examples/seqpatcher/assemblies -o Results -t mmf.csv -O sanger.fasta -g 10 -3 True -x del`. It will generate consensus sequences from the Sanger trace files and integrate them into the ngs consensus. The mmf.csv file tells the application whether the sanger data were paired or single ab1, or in single fasta, and sanger.fasta containing all sanger sequences.
 
-# Features
-
-`codoncounter` and `nucsub` commands were specifically defined for this program. However, each sub-command can be installed and used independently of this program.
-
 # Warning
 
 Recursive use of newly generated consensus sequences might result in incorrect final sequence due to the integration indel events.
 
 # Bug reporting
 
-For bug reporting, support or to propose a new feature, please open an issue.
+To report a bug, request support or propose a new feature, please open an issue.
 
 # Licence
 
