+++ seqPanther/CodonCounter/subs.py
@@ -4,6 +4,7 @@ import pandas as pd
 import numpy as np
 
 from .codon_table import codon_table
+from copy import deepcopy
 from Bio import Seq
 
 
@@ -25,8 +26,8 @@ def sub_table(coordinates_with_changes, params):
     ) in gff_data.iterrows():  # TODO: Change it to itertuple for large gff
         selected_coordinates = keys & set(range(row["start"], row["end"] + 2))
         for selected_coordinate in selected_coordinates:
-            coordinates_with_change = coordinates_with_changes[
-                selected_coordinate].copy()
+            coordinates_with_change = deepcopy(
+                coordinates_with_changes[selected_coordinate])
             coordinates_with_change["start"] = row["start"]
             coordinates_with_change["end"] = row["end"]
             coordinates_with_change["strand"] = row["strand"]
@@ -39,18 +40,14 @@ def sub_table(coordinates_with_changes, params):
             codon_counts = {}
             bases = coordinates_with_change["bases"]
 
-            if selected_coordinate == 28460:
-                print(bases)
-            for base in bases.keys():
+            for base in bases:
 
                 codons = bases[base]["codon_count"]
                 local_codons = {}
 
-                for extended_codon in codons.keys():
+                for extended_codon in codons:
                     codon = extended_codon[2 - shift:5 - shift]
-                    if selected_coordinate == 28460:
-                        print(extended_codon, codon, shift, codons)
-                    if '-' in codon:
+                    if ('-' in codon) or ('N' in codon):
                         continue
                     if codon not in local_codons:
                         local_codons[codon] = codons[extended_codon]
@@ -75,7 +72,7 @@ def sub_table(coordinates_with_changes, params):
                 if codon_counts[codon] / total_codon_count < alt_codon_frac:
                     codons_to_delete.append(codon)
 
-            for base in list(bases.keys()):
+            for base in bases:
                 for codon in codons_to_delete:
                     try:
                         del coordinates_with_change["bases"][base][
@@ -121,7 +118,7 @@ def sub_table(coordinates_with_changes, params):
 
             coordinates_with_change["amino_pos"] = amino_pos
             coordinates_with_change_cds[(
-                row['start'], row['end'], selected_coordinate,
+                selected_coordinate, row['start'], row['end'],
                 row['strand'])] = coordinates_with_change
 
     # NOTE: Fix below
@@ -133,9 +130,8 @@ def sub_table(coordinates_with_changes, params):
         "alt": [],
         "total": [],
         "coor": [],
-        "start": [],
-        "end": [],
-        "strand": [],
+        "coor_full": [],
+        "gene_range": [],
         "ref_codon": [],
         "ref_codon_count": [],
     }
@@ -147,74 +143,74 @@ def sub_table(coordinates_with_changes, params):
         'nucs_count': []
     }
     # coors_to_delete = []
-    for coor in coordinates_with_change:
-        if 'ref_base' not in coordinates_with_change[coor]:
+    for coor in coordinates_with_change_cds:
+        if 'ref_base' not in coordinates_with_change_cds[coor]:
             # coors_to_delete.append(coor)
             continue
-        bases = coordinates_with_change[coor]["bases"]
+        bases = coordinates_with_change_cds[coor]["bases"]
         nucs = []
         nucs_count = []
         for base in bases:
-            if bases[base]["nuc_count"] > coordinates_with_change[coor][
+            if bases[base]["nuc_count"] > coordinates_with_change_cds[coor][
                     'read_count'] * alt_nuc_frac:
                 nucs.append(base)
[…]
