### 2023-03-15 #5129 @kellyrowland
@jsan4christ thanks for the update.

Having references with just URLs is fine if they don't have a DOI; we just ask that authors include DOIs for references which do have them.

The `docs/CONTRIBUTING.md` file looks like it could use some changes to be made more specific to the code repository. For example, you might add a sentence or two about creating good issues or pull requests. Right now, the file talks about issues that contributors might work on, but the Github repo doesn't have any open issues. I would suggest that you change the contributing guidelines to be a little more general. The "Bug Reporting" section of the `README.md` that you have is a good start, you could think about moving that to `CONTRIBUTING.md` and maybe expand upon what folks should include in issues that they open.
