+++ README.md
@@ -10,7 +10,7 @@ SeqPanther consists of the following set of commands:
 
 - **cc2ns** (codon counter 2 nucleotide substitution): performs transformation of codon counter output in a format where the user can select a variant to integrate into the reference or assemblies using the **cc2ns** command. The user is encourage to inspect and edit the output from **codoncounter** to only keep the desired changes before running the **cc2ns** command.
 
-- **cc2ns**: accepts as input the changes file generated by **cc2ns** and modifies the consensus sequences accordingly i.e. adds or changes the mutations as specified in the file. <!--Providing options to users to select changes of their interests-->
+- **nucsubs**: accepts as input the changes file generated by **cc2ns** and modifies the consensus sequences accordingly i.e. adds or changes the mutations as specified in the file. <!--Providing options to users to select changes of their interests-->
 
 - **seqpatcher**: integrates sanger sequencing of missing regions of an incomplete assembly to the assembly. This command is a modification of the [SeqPatcher tool](https://github.com/krisp-kwazulu-natal/seqPatcher).
 
@@ -75,9 +75,9 @@ This command help is accessible using `seqpanther codoncounter` or `seqpanther c
 
 This command help is accessible using `seqpanther cc2ns` or `seqpanther cc2ns --help`.
 
-## nucsub
+## nucsubs
 
-This command help is accessible using `seqpanther nucsub` or `seqpanther nucsub --help`.
+This command help is accessible using `seqpanther nucsubs` or `seqpanther nucsubs --help`.
 
 ## SeqPatcher
 
