### 2023-04-13 #5305 @cbrueffer
On point `Contribution and authorship`, looking at the git history the submitting author does not seem to have made any contribution to the software itself, only the paper. However, from the first commit it seems that the software was developed outside of the present repository, so it's possible that more direct software contributions were made there, but I cannot check that.
