+++ docs/CONTRIBUTING.md
@@ -1,5 +1,32 @@
-#Contribute to codemeleon/seqPanther
+#Contribute to SeqPanther
 
-Make your first contribution to this repository by tackling one of the issues listed below.
+## Report a Bug 
 
-Each issue displayed here is a "good first issue," selected for its relative approachability for first-time contributors.
+If you encounter a bug with while installing or using SeqPanther, please let us know by describing the issue you are facing, reproduction steps, any logs you might have, and our team will address it as quickly as possible.
+
+To report a bug please open an issue.
+
+
+## Request Support or Propose a New Feature
+
+SeqPanther was originally developed for internal use within the (Tulio Lab)[http://krisp.org.za]. We anticipate that usage scenarios will change and additional features required to serve the wider audience.
+
+To request support or propose a new feature, please open an issue or reach out to us via email: sanemmanueljames [at] gmail [dot] org, anmol.kiran [at] gmail [dot] com or tuliodna [at] gmail [dot] com.
+
+
+## Submit a Pull Request
+
+If you have modified SeqPanther to improve its functionality or address any of the open issues, send us a pull request. Our team will review and integrate the changes you have made. 
+
+Make your first contribution to this repository by tackling one of the issues listed below. Each issue displayed here is a "good first issue," selected for its relative approachability for first-time contributors.
+
+## Submitting a Good Pull Request / Issue
+
+A good pull request is one that you understand the point of, clarifies the intentions of the code changes, summarizes the code well and identifies some of the thought process involved in writing the code. Also, a great pull request answers some questions before they are asked. Ensure that you provide as much detail as possible in the description. If your code carries a risk such as a security vulnerability, slower performance or fail under certain conditions, be sure to let us know. Finally, provide a testing plan to help us test and integrate your code smoothly.
+
+If this is your first pull request, refer to the (pull request template)[https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/creating-a-pull-request-template-for-your-repository] to help you prepare the submission. Find out more [here](https://www.pullrequest.com/blog/writing-a-great-pull-request-description/).
+
+Similarly, a good issue will enable you get more help and faster by making it easy for outside contributors to understand your project and help you out.
+
+Ensure that your issue clearly describes what is going on, includes relevant user research, points to the necessary files, and has a checklist of actions to take. Learn more (here)[https://github.com/codeforamerica/howto/blob/master/Good-GitHub-Issues.md] about how to create a good issue.
+
