+++ seqPanther/CodonCounter/indel_frames.py
@@ -10,6 +10,7 @@ def indel_frames(indel_pos_type_size, params):
     alt_codon_frac = params["alt_codon_frac"]
     rid = params["rid"]
     sample = params["sample"]
+    indels_changes = indel_pos_type_size.copy()
     coors = set(indel_pos_type_size["coor"])
     indel_pos_type_size["amino_pos"] = 0
     indel_pos_type_size["codon_pos"] = 0
@@ -18,15 +19,12 @@ def indel_frames(indel_pos_type_size, params):
     indel_pos_type_size['gene_range'] = '-'
     shift, r_shift = 0, 0
     indel_pos_type_size_fragmented = []
-    print(indel_pos_type_size[indel_pos_type_size['count'] > 10])
+    print(indel_pos_type_size[indel_pos_type_size['count'] > 10], "Kiran")
 
     for coor in coors:
         # TODO: use df.to_dict('records'). more detail https://towardsdatascience.com/heres-the-most-efficient-way-to-iterate-through-your-pandas-dataframe-4dad88ac92ee
-        t_gff_data = gff_data[
-            (gff_data["start"] <= coor)
-            & (gff_data["end"] > coor)].to_dict(
-                'record'
-            )  # TODO: Need to make some correction when indel contain end site of CDS or stop codon
+        t_gff_data = gff_data[(gff_data["start"] <= coor)
+                              & (gff_data["end"] >= coor)].to_dict('record')
 
         if not len(t_gff_data):
             print(f'No CDS gff data found for {coor}')
@@ -34,24 +32,45 @@ def indel_frames(indel_pos_type_size, params):
         for gff_row in t_gff_data:
             t_indel_pos_type_size = indel_pos_type_size[
                 indel_pos_type_size["coor"] == coor]
-            adjusted_coor = coor + 1
-            shift = (adjusted_coor - gff_row["start"]) % 3  # + 1
-            r_shift = (3 - shift) % 3
+            shift = (coor - gff_row["start"] +
+                     1) % 3  # original reference is zero based
+
+            t_indel_pos_type_size.loc[t_indel_pos_type_size["coor"] == coor,
+                                      "shift"] = shift
+            # t_indel_pos_type_size.loc[(t_indel_pos_type_size["coor"] == coor) &
+            # (t_indel_pos_type_size["indel"] < 0),
+            # "shift"] += 1
+            t_indel_pos_type_size.loc[(
+                t_indel_pos_type_size["coor"] == coor
+            ), "r_shift"] = (3 - t_indel_pos_type_size.loc[
+                (t_indel_pos_type_size["coor"] == coor), "shift"].values) % 3
             t_indel_pos_type_size.loc[
                 t_indel_pos_type_size["coor"] == coor,
-                ["ref", "read"]] = t_indel_pos_type_size.loc[
+                ["shift", "r_shift"]] = t_indel_pos_type_size.loc[
                     t_indel_pos_type_size["coor"] == coor,
-                    ["ref", "read"]].applymap(
-                        lambda x: x[3 - shift:-(3 - r_shift)])
-            if coor == 1444:
-                print(t_indel_pos_type_size)
-            t_indel_pos_type_size.loc[t_indel_pos_type_size["coor"] == coor,
-                                      ["shift", "r_shift"]] = [shift, r_shift]
+                    ["shift", "r_shift"]].applymap(lambda x: 3 - x)
 
-            amino_pos = (adjusted_coor -
-                         gff_row["start"]) // 3  # - 1 * row["indel"]
[…]
