### 2023-05-08 #5305 @kellyrowland
Hi @cbrueffer @jsan4christ thanks for getting started on this. I see that https://github.com/codemeleon/seqPanther/pull/4 was opened 3 weeks ago - could this be reviewed and merged so that the review here can proceed?
