### 2023-05-12 #5305 @jsan4christ
Thanks Kelly,

I already merged in the pull request.

On Mon, May 8, 2023 at 12:35 PM Kelly L. Rowland ***@***.***>
wrote:

> Hi @cbrueffer <https://github.com/cbrueffer> @jsan4christ
> <https://github.com/jsan4christ> thanks for getting started on this. I
> see that codemeleon/seqPanther#4
> <https://github.com/codemeleon/seqPanther/pull/4> was opened 3 weeks ago
> - could this be reviewed and merged so that the review here can proceed?
>
