### 2023-05-12 #5305 @cbrueffer
Hi @jsan4christ, thanks for merging. There are two more PRs open in the repository that I opened in March you could have a look at. Another will come this weekend and I'll make sure to finish the review until then (sorry for the delay!).

Following the review checklist, there are some points I can mention at this point that could/should be improved:

- software release: while the software has a version (0.0.1), no tag and/or release has been made. This is a problem for reproducibility.
- installation: with the software release above done, you can submit your software to a repository such as Bioconda, which you already use for your dependencies. That would ease installation. I'm happy to help get your software merged in Bioconda.
- testing: right now there are no automated tests. Some functionality tests would be great, but at the very least a basic test that installs the software and dependencies, and checks that it can be executed / prints the usage info would be good.
