+++ seqPanther/CodonCounter/indel_frames.py
@@ -2,7 +2,7 @@
 
 import pandas as pd
 
-from Bio import Seq
+from Bio.Seq import Seq
 
 
 def indel_frames(indel_pos_type_size, params):
@@ -12,19 +12,13 @@ def indel_frames(indel_pos_type_size, params):
     sample = params["sample"]
     indels_changes = indel_pos_type_size.copy()
     coors = set(indel_pos_type_size["coor"])
-    indel_pos_type_size["amino_pos"] = 0
-    indel_pos_type_size["codon_pos"] = 0
-    indel_pos_type_size["shift"] = 0
-    indel_pos_type_size["r_shift"] = 0
-    indel_pos_type_size['gene_range'] = '-'
-    shift, r_shift = 0, 0
     indel_pos_type_size_fragmented = []
-    print(indel_pos_type_size[indel_pos_type_size['count'] > 10], "Kiran")
+    indel_pos_type_size["r_shift"] = 0
 
     for coor in coors:
         # TODO: use df.to_dict('records'). more detail https://towardsdatascience.com/heres-the-most-efficient-way-to-iterate-through-your-pandas-dataframe-4dad88ac92ee
         t_gff_data = gff_data[(gff_data["start"] <= coor)
-                              & (gff_data["end"] >= coor)].to_dict('record')
+                              & (gff_data["end"] >= coor)].to_dict('records')
 
         if not len(t_gff_data):
             print(f'No CDS gff data found for {coor}')
@@ -32,11 +26,16 @@ def indel_frames(indel_pos_type_size, params):
         for gff_row in t_gff_data:
             t_indel_pos_type_size = indel_pos_type_size[
                 indel_pos_type_size["coor"] == coor]
+
             shift = (coor - gff_row["start"] +
                      1) % 3  # original reference is zero based
 
-            t_indel_pos_type_size.loc[t_indel_pos_type_size["coor"] == coor,
-                                      "shift"] = shift
+            t_indel_pos_type_size["shift"] = shift
+            # t_indel_pos_type_size.loc[t_indel_pos_type_size.indel < 0,
+            # "shift"] = t_indel_pos_type_size.loc[
+            # t_indel_pos_type_size.indel < 0,
+            # "shift"].apply(lambda x: (x + 2) % 3)
+
             # t_indel_pos_type_size.loc[(t_indel_pos_type_size["coor"] == coor) &
             # (t_indel_pos_type_size["indel"] < 0),
             # "shift"] += 1
@@ -50,44 +49,12 @@ def indel_frames(indel_pos_type_size, params):
                     t_indel_pos_type_size["coor"] == coor,
                     ["shift", "r_shift"]].applymap(lambda x: 3 - x)
 
-            # if t_indel_pos_type_size['indel'] < 0:
-            # shift = (shift + 1) % 3
-
-            # r_shift = (3 - shift)
-            # codon_pos = coor - shift
-            # shift, r_shift = (shift, r_shift) if shift else (3, 3)
-            # t_indel_pos_type_size.loc[
-            # t_indel_pos_type_size["coor"] == coor,
-            # ["ref", "read"]] = t_indel_pos_type_size.loc[
-            # t_indel_pos_type_size["coor"] == coor,
-            # ["ref", "read"]].applymap(
-            # lambda x: x[3 - shift:-(3 - r_shift)])
-            # if coor == 1444:
-            # print(t_indel_pos_type_size)
-            # t_indel_pos_type_size.loc[t_indel_pos_type_size["coor"] == coor,
-            # ["shift", "r_shift"]] = [shift, r_shift]
-
-            amino_pos = (coor - gff_row["start"]) // 3
[…]
