+++ seqPanther/CodonCounter/bammer.py
@@ -4,11 +4,20 @@ import pysam
 from os import path
 
 
-def check_sort_and_index_bam(bam_file, tmp_dir):
+def check_sort_and_index_bam(bam_file, rid, tmp_dir):
     """
-    Checks file is sorted or not. If not, sorts and indexes.
+    Checks bam file has given reference id.
+    It also checks file is sorted or not. If not, sorts and indexes.
     """
-    header = pysam.AlignmentFile(bam_file, 'rb').header.as_dict()
+    samfile = pysam.AlignmentFile(bam_file, 'rb')
+    header = samfile.header.as_dict()
+    if rid not in samfile.references:
+        print(f"Given reference {rid} not in given bam file {bam_file}")
+        print("List of references")
+        print(samfile.references)
+        print(f"Ignoring {bam_file}")
+        return
+
     if header['HD']['SO'] == 'coordinate':
         if not path.exists(bam_file + ".bai"):
             pysam.index(bam_file)
