### 2023-05-15 codemeleon/seqPanther#10 @cbrueffer
Adding guidelines for prospective contributors
Some guidelines for prospective contributors would be useful, as highlighted also in the JOSS review checklist. A good example is Biopython, where information is mentioned both in the project README, and in more detail in https://github.com/biopython/biopython/blob/master/CONTRIBUTING.rst

This issue is part of the JOSS review at https://github.com/openjournals/joss-reviews/issues/5305
