### 2023-05-21 codemeleon/seqPanther#11 @cbrueffer
Manuscript: reference issue
There seems to be an issue with one citation in the manuscript, `molecular clock 19 and`, not sure which reference is intended here.

https://github.com/codemeleon/seqPanther/blob/master/manuscript/paper.md?plain=1#LL88C1394-L88C1394

This issue is part of the JOSS review at https://github.com/openjournals/joss-reviews/issues/5305
