### 2023-05-26 #5129 @jsan4christ
> @jsan4christ below are some points that need your attention:
>
> * [ ]  On the above paper compilation error you need to use quotes if you have multiple affiliations, e.g. `1,2` should be written `"1,2"`. I have fixed this for you here: [Added quotes for affiliations in JOSS paper draft codemeleon/seqPanther#1](https://github.com/codemeleon/seqPanther/pull/1). You can call `@editorialbot generate pdf` here to update the paper draft.
> * [ ]  Can  you check/fix those potentially missing DOIs? ☝️ When you do, you can call `@editorial bot check references` here, to see if they are fixed.
> * [ ]  During the review process the editor/reviewer will also look for contributing guidelines. These appear missing for your project. Consider adding a `CONTRIBUTING.md` file to your main folder and to explain how folks can contribute to your project. More information in this GitHub guide: https://docs.github.com/articles/setting-guidelines-for-repository-contributors/

Thank you for these important comments;

x For point one, I just merged your pool request. Thanks for submitting it.

x For the dois, we have added all the dois, except where they are not applicable such in the case of web-pages or unpublished material.

x We added a contribution.md section and it's a living document, meaning, we will continue to develop this section as time goes on and as we interact more with potential contributors.
