### 2023-06-02 codemeleon/seqPanther#11 @jsan4christ
Thanks for picking this up.

It is:

Natural selection on synonymous mutations in SARS-CoV-2 and the impact on estimating divergence time (https://doi.org/10.2217/fvl-2021-0078)

---
Estimation of the genuine divergence time based on synonymous sites
We have shown that the synonymous mutations in SARS-CoV-2 do not occur under constant rate, and that they also undergo natural selection. Both features hinder the accurate estimation of divergence time based on synonymous sites.
---

I have corrected accordingly.
