### 2023-06-02 47ccffc
Resolve issue #11
Corrected citation to [@Yu_2021]
