+++ manuscript/paper.md
@@ -85,7 +85,7 @@ The Codoncounter module takes at least a BAM file, reference FASTA and GFF genom
 
 ![Visual illustration of the ``SeqPanther`` components. A) codoncounter which accepts a BAM file or directory containing BAM files (if running batch mode), a reference sequence and a genomic region or set of positions and generates two reports (codon and nucleotide) and a genome coverage and depth map. B) The cc2ns command accepts indel and substitution reports as inputs and generates a substitution file containing all the nucleotide substitutions and indels as well as their frequency. C) The nucsubs module that takes the substitutions file, reviewed and edited by the user and modifies the consensus sequences. D) Depth and coverage plot generated by codoncounter. Colored circles represent different types of alterations across the length of the assembly, which are supported by more than 5% of total reads.\label{fig:Figure1}](Figure1.jpg)
 
-Additionally, for each output type generated by Codoncounter, the command provides additional statistics. These include the total number of reads, reads mapping to the reference, alternate substitutions and the percentage they comprise. Typically, the codon results are tied to a user-specified region such as an open reading frame (ORF) or protein-coding gene with coordinates defined in a general feature format (GFF) file. The focus on a user-specified genomic segment is based on the fact that some samples might have an extremely large mutation load and quality control will only be focused on just a small subset of mutations in a region of interest, for example, in the case of SARS-CoV-2 mutations in the Spike gene or the Receptor Binding Domain (RBD). The user can also specify the set of nucleotide positions or a genomic range of interest or both. The user-provided reference sequence ID must match the ID in the BAM, GFF, and reference FASTA file. The codons report only contains non-synonymous substitutions or 3x nucleotide indels in the mapped reads (Supplementary Table 1). In comparison, all nucleotide changes are reported including the synonymous changes i.e. mutations not associated with an amino acid change (Supplementary Table 2). The report is thus also useful to investigate synonymous mutations which, although not changing the encoded amino acid, play a crucial role in the estimation of the molecular clock 19 and designation of strains and lineages. Examples of lineage-defining synonymous mutations include the T23341C and T24187A in the P320 variant and C26801T in the Alpha/B.1.1.7 variant [@covlin_2020].
+Additionally, for each output type generated by Codoncounter, the command provides additional statistics. These include the total number of reads, reads mapping to the reference, alternate substitutions and the percentage they comprise. Typically, the codon results are tied to a user-specified region such as an open reading frame (ORF) or protein-coding gene with coordinates defined in a general feature format (GFF) file. The focus on a user-specified genomic segment is based on the fact that some samples might have an extremely large mutation load and quality control will only be focused on just a small subset of mutations in a region of interest, for example, in the case of SARS-CoV-2 mutations in the Spike gene or the Receptor Binding Domain (RBD). The user can also specify the set of nucleotide positions or a genomic range of interest or both. The user-provided reference sequence ID must match the ID in the BAM, GFF, and reference FASTA file. The codons report only contains non-synonymous substitutions or 3x nucleotide indels in the mapped reads (Supplementary Table 1). In comparison, all nucleotide changes are reported including the synonymous changes i.e. mutations not associated with an amino acid change (Supplementary Table 2). The report is thus also useful to investigate synonymous mutations which, although not changing the encoded amino acid, play a crucial role in the estimation of the molecular clock [@Yu_2021] and designation of strains and lineages. Examples of lineage-defining synonymous mutations include the T23341C and T24187A in the P320 variant and C26801T in the Alpha/B.1.1.7 variant [@covlin_2020].
 
 Table 4: Types of changes resulting in amino acid alterations. In the case of indels, only multiples of 3 consecutive nucleotide indels were considered for integration in consensus sequences.
 
