### 2023-06-10 #5305 @jsan4christ
@cbrueffer can we begin to make some progress on this?:  installation: with the software release above done, you can submit your software to a repository such as Bioconda, which you already use for your dependencies. That would ease installation. I'm happy to help get your software merged in Bioconda.
