### 2023-06-11 1e232d9
environment file for conda users
