+++ CONTRIBUTING.md
@@ -1,0 +1,24 @@
+# Contributing to SeqPanther
+
+We encourage the submission of pull requests for bug fixes or the addition of new features. Kindly share or integrate your code, as it can be beneficial for other users.
+
+# Coding Conventions
+
+SeqPanther tries to follow the coding conventions laid out in PEP8 and PEP257.
+
+- http://www.python.org/dev/peps/pep-0008/
+- http://www.python.org/dev/peps/pep-0257/
+
+To facilitate style checking, we make use of `black` formatter on the Python code.
+
+# Testing
+
+Tests are required for any new feature or functionality to be considered. Similarly, when addressing a bug fix, we strongly recommend including input data, code, and expected outcomes.
+
+# Continuous Integration
+
+Upon submitting your pull request on GitHub, we will conduct thorough testing. Following validation of the input, code, and expected outputs, the provided code will be merged into the repository.
+
+# Contributing to the SeqPanther tutorial
+
+Please add suggestions to [README.md](https://github.com/codemeleon/seqPanther/blob/master/README.md) or open an issue.
