+++ seqPanther/CodonCounter/subs.py
@@ -1,133 +1,112 @@
 #!/usr/bin/env python
 
 import pandas as pd
+import numpy as np
+
 from .codon_table import codon_table
-from os import path
+from copy import deepcopy
 from Bio import Seq
-import pysam
 
 
-def sub_table(coordinates_with_change, bam, params):
-    sample = path.split(bam)[1]
+def sub_table(coordinates_with_changes, params):
     sequences = params["sequences"]
     rid = params["rid"]
-    ignore_orphans = params["ignore_orphans"]
-    ignore_overlaps = params["ignore_overlaps"]
-    min_base_quality = params["min_base_quality"]
+    sample = params["sample"]
     alt_codon_frac = params["alt_codon_frac"]
+    alt_nuc_frac = params["alt_nuc_count"]
 
     gff_data = params["gff_data"]
     min_seq_depth = params["min_seq_depth"]
-    sample = path.split(bam)[1]
-    keys = set(coordinates_with_change)
-    samfile = pysam.AlignmentFile(bam, "rb")
+    keys = set(coordinates_with_changes)
+    coordinates_with_change_cds = {}
 
     for (
             _,
             row,
     ) in gff_data.iterrows():  # TODO: Change it to itertuple for large gff
-        selected_coordinates = keys & set(range(row["start"], row["end"] + 1))
+        selected_coordinates = keys & set(range(row["start"], row["end"] + 2))
         for selected_coordinate in selected_coordinates:
-            coordinates_with_change[selected_coordinate]["start"] = row[
-                "start"]
-            coordinates_with_change[selected_coordinate]["end"] = row["end"]
-            coordinates_with_change[selected_coordinate]["strand"] = row[
-                "strand"]
+            coordinates_with_change = deepcopy(
+                coordinates_with_changes[selected_coordinate])
+            coordinates_with_change["start"] = row["start"]
+            coordinates_with_change["end"] = row["end"]
+            coordinates_with_change["strand"] = row["strand"]
             shift = (selected_coordinate - row["start"]) % 3
-            iter = samfile.pileup(
-                rid,
-                selected_coordinate,
-                selected_coordinate + 1,
-                ignore_orphans=ignore_orphans,
-                min_mapping_quality=min_base_quality,
-                min_base_quality=min_base_quality,
-                ignore_overlaps=ignore_overlaps,
-                max_depth=1000000,
-            )
             ref_base = sequences[selected_coordinate].seq
             ref_codon = sequences[selected_coordinate -
                                   shift:selected_coordinate - shift + 3].seq
-            total_codon_count = 0
             ref_codon_count = 0
-            for pileupcol in iter:
-                if pileupcol.pos != selected_coordinate:
-                    continue
-                for pread in pileupcol.pileups:
-                    if not pread.is_del and not pread.is_refskip:
-                        codon = pread.alignment.query_sequence[
-                            pread.query_position - shift:pread.query_position -
-                            shift + 3]
-                        if codon == ref_codon:
-                            ref_codon_count += 1
-                        if (codon in codon_table) and (
-                                pread.alignment.query_sequence[
-                                    pread.query_position]
-                                in coordinates_with_change[
-                                    pileupcol.pos]["bases"]):
+            total_codon_count = 0
+            codon_counts = {}
+            bases = coordinates_with_change["bases"]
 
-                            if (codon not in coordinates_with_change[
-                                    pileupcol.pos]["bases"]
-                                [pread.alignment.query_sequence[
-                                    pread.query_position]]["codon_count"]):
-                                coordinates_with_change[pileupcol.pos][
[…]
