+++ seqPanther/CodonCounter/indel_frames.py
@@ -1,303 +1,197 @@
 #!/usr/bin/env python
 
-import pysam
 import pandas as pd
-from collections import Counter
 
-from Bio import Seq
-import pyfaidx
-from os import path
+from Bio.Seq import Seq
 
 
 def indel_frames(indel_pos_type_size, bam, params):
-    rid = params["rid"]
     gff_data = params["gff_data"]
-    sequences = pyfaidx.Fasta(params["ref"])[rid]
-    ignore_orphans = params["ignore_orphans"]
-    min_mapping_quality = params["min_mapping_quality"]
-    min_base_quality = params["min_base_quality"]
-    ignore_overlaps = params["ignore_overlaps"]
-    alt_nuc_count = params["alt_nuc_count"]
     alt_codon_frac = params["alt_codon_frac"]
-    sample = path.split(bam)[1].split(".bam")[0]
+    rid = params["rid"]
+    sample = params["sample"]
+    coors = set(indel_pos_type_size["coor"])
+    indel_pos_type_size_fragmented = []
+    indel_pos_type_size["r_shift"] = 0
 
-    deletion_frame = {}
-    insertion_frame = {}
-    samfile = pysam.AlignmentFile(bam, "rb")
+    for coor in coors:
+        # TODO: use df.to_dict('records'). more detail https://towardsdatascience.com/heres-the-most-efficient-way-to-iterate-through-your-pandas-dataframe-4dad88ac92ee
+        t_gff_data = gff_data[(gff_data["start"] <= coor)
+                              & (gff_data["end"] >= coor)].to_dict('records')
 
-    for _, row in indel_pos_type_size.iterrows():
-        t_gff_data = gff_data[(gff_data["start"] <= row["coor"])
-                              & (gff_data["end"] > row["coor"])]
         if not len(t_gff_data):
-            print(f'No gff data found for {row["coor"]}')
+            print(f'No CDS gff data found for position {coor} in file {bam}.')
             continue
-        for _, gff_row in t_gff_data.iterrows():
-            iter = samfile.pileup(
-                rid,
-                row["coor"],
-                row["coor"] + 1,
-                ignore_orphans=ignore_orphans,
-                min_mapping_quality=min_mapping_quality,
-                min_base_quality=min_base_quality,
-                ignore_overlaps=ignore_overlaps,
-                max_depth=1000000,
-            )
+        for gff_row in t_gff_data:
+            t_indel_pos_type_size = indel_pos_type_size[
+                indel_pos_type_size["coor"] == coor]
 
-            if row["indel"] < 0:
-                for pileupcol in iter:
-                    if pileupcol.pos != row["coor"]:
-                        continue
-                    adjusted_coor = row["coor"] + 1
-                    shift = (adjusted_coor - gff_row["start"]) % 3  # + 1
-                    r_shift = (3 - shift) % 3
-                    ref_sub_seq = sequences[adjusted_coor -
-                                            shift:adjusted_coor -
-                                            row["indel"] + r_shift].seq
-                    amino_pos = (adjusted_coor -
-                                 gff_row["start"]) // 3  # - 1 * row["indel"]
-                    ref_count = 0
-                    depth = 0
-                    deleted_codon = []
-                    for pread in pileupcol.pileups:
-                        if (pread.alignment.reference_start <= pileupcol.pos
-                                and pileupcol.pos - row["indel"] <
-                                pread.alignment.reference_end):
-                            depth += 1
-                        if pread.indel:
-                            if (pread.indel < 0) and (pread.indel
-                                                      == row["indel"]):
-                                read_sub_seq = pread.alignment.query_sequence[
-                                    pread.query_position - shift +
[…]
