### 2023-06-11 codemeleon/seqPanther#10 @jsan4christ
Thanks @cbrueffer.

This has been addressed in commit [5ccc42c](https://github.com/codemeleon/seqPanther/commit/5ccc42c09d3e9087867caeb7351bdcc11d43d7ae).
