### 2023-06-22 #5305 @cbrueffer
@jsan4christ Yes, the tests are the only outstanding functional issue in my view. The other remaining one is `contribution and authorship` in the checklist, see below.

@kellyrowland @ctb the item *Contribution and authorship* may need an editorial decision / guidance, I'm not sure I can tick it off as is (see also my comments here from April 13), but this may be subject to interpretation and I'm happy to be convinced otherwise.

The checklist states `Has the submitting author (@jsan4christ) made major contributions to the software?` This depends on what exactly "software" here entails, but judging only from the repository at https://github.com/codemeleon/seqPanther/, and that's the only thing I have to assess this, I'd have to say no. In that repository the submitting author contributed to the paper and software-adjacent files (README etc, also as part of this review), but as far as I see made no contributions to the actual software code. I cannot judge other contributions (e.g. ideas etc) or whether there was a repository that predates the current, where code contributions happened. I also cannot judge the contribution of the other authors on the paper, other than the last author, who appears to have written the software code at hand.

@ctb how do you view this, I see you ticked this item off? @kellyrowland Any guidance on how to proceed here?
