### 2023-06-26 4577116
minor corrections & updates to README
