+++ README.md
@@ -1,6 +1,6 @@
 # SeqPanther
 
-`SeqPanther` is a Python application that provides the user with a suite of tools to further interrogate the circumstance under which certain mutations occur or are missed and to enables the user to modify the consensus as needed. The tool is applicable to non-segmented bacterial and viral genomes where reads are mapped to a reference sequence. `SeqPanther` generates detailed reports of mutations identified within a genomic segment or positions of interest, including visualization of the genome coverage and depth. The tool is particularly useful in the examination of multiple next-generation sequencing (NGS) short-read samples. Additionally, we have integrated `Seqpatcher`[singh et. al.](https://www.microbiologyresearch.org/content/journal/mgen/10.1099/mgen.0.000774) which supports the merging of Sanger sequences, or consensus thereof into their respective NGS consensus.
+`SeqPanther` is a Python application that provides the user with a suite of tools to further interrogate the circumstance under which certain mutations occur or are missed and to enables the user to modify the consensus as needed. The tool is applicable to non-segmented bacterial and viral genomes where reads are mapped to a reference sequence. `SeqPanther` generates detailed reports of mutations identified within a genomic segment or positions of interest, including visualization of the genome coverage and depth. The tool is particularly useful in the examination of multiple next-generation sequencing (NGS) short-read samples. Additionally, we have integrated `Seqpatcher` [Singh et. al.](https://www.microbiologyresearch.org/content/journal/mgen/10.1099/mgen.0.000774) which supports the merging of Sanger sequences, or consensus thereof into their respective NGS consensus.
 
 ![Seqpanther workflow, genome coverage plot and codon count report](https://github.com/codemeleon/seqPanther/blob/master/manuscript/banner.jpg)
 
@@ -16,11 +16,11 @@ SeqPanther consists of the following set of commands:
 
 ## Operating system compatibility
 
-Unix and OSX Commandline Application. 
+Unix and OS X Commandline Application. 
 
 ## Dependencies
 
-The tool relies on multiple external opensource programs and python modules as listed below:
+The tool relies on multiple external open source programs and python modules as listed below:
 
 ### External Tools
 
@@ -89,7 +89,7 @@ This command help is accessible at `seqpanther seqpatcher` or `seqpanther seqpat
 
 1. A sample dataset of five BAM files from South African SARS-CoV-2 samples can be downloaded from [SeqPanther zenodo](https://zenodo.org/record/7601513) page. The BAM files were generated by mapping reads mapped against reference (NC_045512.2) using [BWA](https://github.com/lh3/bwa). BWA was also used to generate the SAM files which were converted to BAM and sorted using samtools.
 
-2. Reference sequences in Fasta format and genome annotation GFF files were downloaded from [NCBI Genome Fasta](https://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/009/858/895/GCF_009858895.2_ASM985889v3/GCF_009858895.2_ASM985889v3_genomic.fna.gz) and [ NCBI Genome GFF](https://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/009/858/895/GCF_009858895.2_ASM985889v3/GCF_009858895.2_ASM985889v3_genomic.gff.gz) respectively. The downloaded files need to un-archived.
[…]
