### 2023-06-26 #5305 @ctb
OK, I ran through installation and tested the examples, ticking off the rest of the checkboxes. I made some notes [here](https://hackmd.io/FJKxrD8yQvCBy8y5wgGmNg?view) in a hackmd. You are welcome to take what you wish from the hackmd!

I also updated the README with some minor corrections (sorry, I proofread by habit 😆 ). Welcome to merge them or not.

I am satisfied with the software and it is ready for acceptance IMO - thanks to @kellyrowland for his more thorough and detailed review ;).

🎉

The main piece of advice I have to the authors is to streamline the tutorial in the README so that the commands are clearly marked & copy-paste-able. This doesn't affect the functionality of the software but might have a big impact on adoption 🤷 .
