### 2023-07-05 #5305 @kellyrowland
Hi @jsan4christ I've opened one minor issue on your repo, but the proof looks good otherwise.

After addressing this, please issue a new tagged release of the software (if changed since the start of the review), and archive it (on [Zenodo](https://zenodo.org/), [figshare](https://figshare.com/), or elsewhere). Then, please then post the version number and archive DOI here in the review issue, and I'll follow some subsequent wrap-up steps from there.
