### 2023-07-10 9d8d8bd
minor fix to matplotlib bibliography entry
semicolon placed in journal title causing JOSS publish steps to fail
