+++ manuscript/paper.bib
@@ -252,7 +252,7 @@
 	pages = {90--95},
 	author = {John D. Hunter},
 	title = {Matplotlib: A 2D Graphics Environment},
-	journal = {Computing in Science {\&}amp$\mathsemicolon$ Engineering}
+	journal = {Computing in Science \& Engineering}
 }
 
 
