+++ manuscript/paper.md
@@ -55,7 +55,7 @@ Keywords: Bioinformatics, sequence analysis, NGS, codon, amino acid substitution
 
 # Abstract
 
-Pathogen genomes harbor critical information necessary to support genomic investigations that inform public health interventions such as treatment, control, and eradication. To extract this information, their sequences are analyzed to identify structural variations such as single nucleotide polymorphisms (SNPs) and insertions and deletions (INDELs) that may be associated with phenotypes of interest. Typically, this involves the assembly of reads, calling of variants, and calculation of a consensus sequence  for downstream analysis. Several pipelines exist to perform quality control (QC) of raw reads and whole-genome assemblies. However, there is no easy-to-use, freely available bioinformatics QC tool to explore mappings for both positional codons and nucleotide distributions in mapped short reads of microbial genomes. To address this problem, we have developed a fast and accurate tool to summarise read counts associated with codons, nucleotides, and INDELs in mapped next-generation sequencing (NGS) short reads. The tool, developed in Python, also provides a visualization of the genome sequencing depth and coverage. Furthermore, the tool can be run in single or batch mode, where several genomes need to be analyzed. Our tool produces a text-based report that enables quick review or can be imported into any analytical tool for upstream analysis. Additionally, the tool provides the functionality to modify consensus sequences by adding, masking, or restoring to wild-type mutations specified by the user.
+Pathogen genomes harbor critical information necessary to support genomic investigations that inform public health interventions such as treatment, control, and eradication. To extract this information, their sequences are analyzed to identify structural variations such as single nucleotide polymorphisms (SNPs) and insertions and deletions (INDELs) that may be associated with phenotypes of interest. Typically, this involves the assembly of reads, calling of variants, and calculation of a consensus sequence for downstream analysis. Several pipelines exist to perform quality control (QC) of raw reads and whole-genome assemblies. However, there is no easy-to-use, freely available bioinformatics QC tool to explore mappings for both positional codons and nucleotide distributions in mapped short reads of microbial genomes. To address this problem, we have developed a fast and accurate tool to summarise read counts associated with codons, nucleotides, and INDELs in mapped next-generation sequencing (NGS) short reads. The tool, developed in Python, also provides a visualization of the genome sequencing depth and coverage. Furthermore, the tool can be run in single or batch mode, where several genomes need to be analyzed. Our tool produces a text-based report that enables quick review or can be imported into any analytical tool for upstream analysis. Additionally, the tool provides the functionality to modify consensus sequences by adding, masking, or restoring to wild-type mutations specified by the user.
 
 Availability: `SeqPanther` is available at https://github.com/codemeleon/seqPanther, along with the necessary documentation for installation and usage.
 
@@ -77,8 +77,8 @@ We utilized Python (v3.9.9) as the base programming language to develop our pipe
 
 # Features
 
-`SeqPanther` is an easy-to-use and accurate command-line tool for generating summary statistics from BAM files relating to structural changes (SNPs and INDELs) relative to a reference sequence and modifying the consensus sequences calculated from the BAM files to add, remove, or change the variations at specified positions within the sequences. ![image](https://github.com/codemeleon/seqPanther/assets/4986950/da9c2982-a68a-47a5-b97f-bff87dc84c4a)
- `SeqPanther` features a suite of tools that perform various functions including codoncounter, cc2ns, and nucsubs. Figure 1 shows the inputs, outputs, and dependencies between the tools.
[…]
