### 2023-07-12 #5305 @kellyrowland
@jsan4christ the bibliography file still has the "\mathsemicolon" entry: https://github.com/codemeleon/seqPanther/blob/master/manuscript/paper.bib#L255

It looks like the pull request here: https://github.com/codemeleon/seqPanther/pull/16 was closed but not merged, so the change hasn't gone into the repo.
