### 2023-07-12 49cc792
Fix figure label.
