+++ manuscript/paper.md
@@ -82,7 +82,7 @@ We utilized Python (v3.9.9) as the base programming language to develop our pipe
 
 ## Codoncounter
 
-The Codoncounter module accepts a BAM file, reference FASTA, and General Feature Format (GFF) genome annotation file as input and produces four output files: 1) a table summarizing nonsynonymous substitutions (Table S1); 2) a table listing the synonymous substitutions (Table S2); 3) a list of the nucleotide INDELs identified (Table S3); and 4) a PDF file with coverage plots and color-coded substitutions at different genomic coordinates that occur relative to the reference \autoref{fig: Figure1}. In the case of unsorted and unindexed BAM files, the command first generates a temporary sorted and indexed BAM file that is used to complete the analysis. Amino acid changes are displayed relative to the strand where the gene is present. A codon-aware alignment of the nucleotides is inferred and used to determine the amino acid changes. Codon changes at INDEL sites are not included in the substitution file but are explored independently in the INDELs file.
+The Codoncounter module accepts a BAM file, reference FASTA, and General Feature Format (GFF) genome annotation file as input and produces four output files: 1) a table summarizing nonsynonymous substitutions (Table S1); 2) a table listing the synonymous substitutions (Table S2); 3) a list of the nucleotide INDELs identified (Table S3); and 4) a PDF file with coverage plots and color-coded substitutions at different genomic coordinates that occur relative to the reference \autoref{fig:Figure1}. In the case of unsorted and unindexed BAM files, the command first generates a temporary sorted and indexed BAM file that is used to complete the analysis. Amino acid changes are displayed relative to the strand where the gene is present. A codon-aware alignment of the nucleotides is inferred and used to determine the amino acid changes. Codon changes at INDEL sites are not included in the substitution file but are explored independently in the INDELs file.
 
 ![Visual illustration of the ``SeqPanther`` components. A) codoncounter accepts a BAM file or directory containing BAM files (if running batch mode), a reference sequence, and a genomic region or set of positions and generates two reports (codon and nucleotide) and a genome coverage and depth map. B) The cc2ns command accepts INDEL(s) and substitution reports as inputs and generates a substitution file containing all the nucleotide substitutions and INDELs as well as their frequency. C) The nucsubs module takes as input the substitutions file, that has been reviewed and edited by the user and modifies the consensus sequences. D) Example depth and coverage plot generated by codoncounter. Colored circles represent different types of alterations across the length of the assembly, which are supported by more than 5% of total reads.\label{fig:Figure1}](Figure1.jpg)
 
