### 2023-07-12 #5305 @cbrueffer
https://github.com/codemeleon/seqPanther/pull/18 may help
