### 2023-07-14 #5305 @Kevin-Mattheus-Moerman
@jsan4christ I am the AEiC for this track and here to help process the final steps. Please look at the above points that require your attention. You need to manually edit the ZENODO archive such that the title matches the paper title, the author set matches the paper authors, and that the license listed matches your software license.
In addition:

- [ ] Please spell out `USA` as `United States of America`.
- [ ] If Affiliation 5 has a "department" to list please add it.
