+++ manuscript/paper.bib
@@ -40,7 +40,7 @@
 	number = {19},
 	pages = {3207--3209},
 	author = {Ka-Kit Lam and Kurt LaButti and Asif Khalak and David Tse},
-	title = {{FinisherSC}: a repeat-aware tool for upgrading$\less$i$\greater$de novo$\less$/i$\greater$assembly using long reads},
+	title = {{FinisherSC}: a repeat-aware tool for upgrading \emph{de novo} assembly using long reads},
 	journal = {Bioinformatics}
 }
 
@@ -56,7 +56,7 @@
 	pages = {e0277659},
 	author = {Sonsiray Alvarez Narvaez and Zhenyu Shen and Lifang Yan and Brianna L. S. Stenger and Laura B. Goodman and Ailam Lim and Ruth H. Nissly and Meera Surendran Nair and Shuping Zhang and Susan Sanchez},
 	editor = {Timothy J. Johnson},
-	title = {Optimized conditions for Listeria, Salmonella and Escherichia whole genome sequencing using the Illumina {iSeq}100 platform with point-and-click bioinformatic analysis},
+	title = {Optimized conditions for Listeria, Salmonella and Escherichia whole genome sequencing using the {Illumina iSeq}100 platform with point-and-click bioinformatic analysis},
 	journal = {{PLOS} {ONE}}
 }
 
@@ -116,7 +116,7 @@
 	volume = {28},
 	number = {6},
 	author = {Melissa Sutton and Tyler S. Radniecki and Devrim Kaya and Dana Alegre and Matthew Geniza and Anne-Marie Girard and Katherine Carter and Mark Dasenko and Justin L. Sanders and Paul R. Cieslak and Christine Kelly and Brett M. Tyler},
-	title = {Detection of {SARS}-{CoV}-2 B.1.351 (Beta) Variant through Wastewater Surveillance before Case Detection in a Community, Oregon, {USA}},
+	title = {Detection of {SARS}-{CoV}-2 {B.1.351 (Beta)} Variant through Wastewater Surveillance before Case Detection in a Community, Oregon, {USA}},
 	journal = {Emerging Infectious Diseases}
 }
 
@@ -154,7 +154,7 @@
 	volume = {820},
 	pages = {153171},
 	author = {Warish Ahmed and Aaron Bivins and Wendy J.M. Smith and Suzanne Metcalfe and Mikayla Stephens and Amy V. Jennison and Frederick A.J. Moore and Jayden Bourke and Sanmarie Schlebusch and Jamie McMahon and Glen Hewitson and Son Nguyen and Jean Barcelon and Greg Jackson and Jochen F. Mueller and John Ehret and Ian Hosegood and Wei Tian and Haofei Wang and Lin Yang and Paul M. Bertsch and Josh Tynan and Kevin V. Thomas and Kyle Bibby and Tyson E. Graber and Ryan Ziels and Stuart L. Simpson},
-	title = {Detection of the Omicron (B.1.1.529) variant of {SARS}-{CoV}-2 in aircraft wastewater},
+	title = {Detection of the {Omicron (B.1.1.529)} variant of {SARS}-{CoV}-2 in aircraft wastewater},
 	journal = {Science of The Total Environment}
 }
 
